[errorhandling] Error Handling 기초 구현 완료 / JPA 연관맵핑 문제해결 요망

+++ backend/src/main/java/com/ssafy/respository/PortfolioImageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.vo.PortfolioImage;
 
 @Repository
-public interface PortfolioImageRepository extends JpaRepository<PortfolioImage, Integer> {
+public interface PortfolioImageRepository extends JpaRepository<PortfolioImage, Long> {
 	
 }
