+++ backend/src/main/java/com/ssafy/service/PortfolioImageService.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import com.ssafy.vo.PortfolioImage;
 
 public interface PortfolioImageService {
-	public Optional<PortfolioImage> findPortfolioImageByPortfolioImageId(final int portfolioImageId); 
+	public Optional<PortfolioImage> findPortfolioImageByPortfolioImageId(final long portfolioImageId);
 	public PortfolioImage savePortfolioImage(final PortfolioImage portfolioImage);
-	public void deletePortfolioImageByPortfolioImageId(final int portfolioImageId); 
-	public int countPortfolioImages(); 
+	public void deletePortfolioImageByPortfolioImageId(final long portfolioImageId); 
+	public long countPortfolioImages(); 
 }
